--- spring-ai-core/src/main/java/org/springframework/ai/rag/analysis/package-info.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2023-2024 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * RAG Module: Query Analysis.
- * <p>
- * This package encompasses all components involved in the pre-retrieval phase of a
- * retrieval augmented generation flow. Queries are transformed, expanded, or constructed
- * so to enhance the effectiveness and accuracy of the subsequent retrieval phase.
- */
-@NonNullApi
-@NonNullFields
-package org.springframework.ai.rag.analysis;
-
-import org.springframework.lang.NonNullApi;
-import org.springframework.lang.NonNullFields;
