--- spring-ai-core/src/main/java/org/springframework/ai/rag/analysis/query/transformation/package-info.java
@@ -1,25 +1,0 @@
-/*
- * Copyright 2023-2024 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * RAG Component: Query Transformation.
- */
-@NonNullApi
-@NonNullFields
-package org.springframework.ai.rag.analysis.query.transformation;
-
-import org.springframework.lang.NonNullApi;
-import org.springframework.lang.NonNullFields;
