+++ spring-ai-core/src/main/java/org/springframework/ai/rag/postretrieval/compression/package-info.java
@@ -15,11 +15,11 @@
  */
 
 /**
- * RAG Sub-Module: Document Search.
+ * RAG Sub-Module: Document Compression.
  */
 @NonNullApi
 @NonNullFields
-package org.springframework.ai.rag.retrieval.search;
+package org.springframework.ai.rag.postretrieval.compression;
 
 import org.springframework.lang.NonNullApi;
 import org.springframework.lang.NonNullFields;
