+++ spring-ai-core/src/main/java/org/springframework/ai/rag/postretrieval/selection/package-info.java
@@ -15,11 +15,11 @@
  */
 
 /**
- * RAG Sub-Module: Document Search.
+ * RAG Sub-Module: Document Selection.
  */
 @NonNullApi
 @NonNullFields
-package org.springframework.ai.rag.retrieval.search;
+package org.springframework.ai.rag.postretrieval.selection;
 
 import org.springframework.lang.NonNullApi;
 import org.springframework.lang.NonNullFields;
