+++ spring-ai-core/src/main/java/org/springframework/ai/rag/preretrieval/query/transformation/package-info.java
@@ -15,11 +15,11 @@
  */
 
 /**
- * RAG Sub-Module: Document Search.
+ * RAG Sub-Module: Query Transformation.
  */
 @NonNullApi
 @NonNullFields
-package org.springframework.ai.rag.retrieval.search;
+package org.springframework.ai.rag.preretrieval.query.transformation;
 
 import org.springframework.lang.NonNullApi;
 import org.springframework.lang.NonNullFields;
