--- spring-ai-core/src/test/java/org/springframework/ai/rag/analysis/query/expansion/MultiQueryExpanderTests.java
@@ -1,72 +1,0 @@
-/*
- * Copyright 2023-2024 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.ai.rag.analysis.query.expansion;
-
-import org.junit.jupiter.api.Test;
-
-import org.springframework.ai.chat.client.ChatClient;
-import org.springframework.ai.chat.prompt.PromptTemplate;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mock;
-
-/**
- * Unit tests for {@link MultiQueryExpander}.
- *
- * @author Thomas Vitale
- */
-class MultiQueryExpanderTests {
-
-	@Test
-	void whenChatClientBuilderIsNullThenThrow() {
-		assertThatThrownBy(() -> MultiQueryExpander.builder().chatClientBuilder(null).build())
-			.isInstanceOf(IllegalArgumentException.class)
-			.hasMessageContaining("chatClientBuilder cannot be null");
-	}
-
-	@Test
-	void whenQueryIsNullThenThrow() {
-		QueryExpander queryExpander = MultiQueryExpander.builder()
-			.chatClientBuilder(mock(ChatClient.Builder.class))
-			.build();
-		assertThatThrownBy(() -> queryExpander.expand(null)).isInstanceOf(IllegalArgumentException.class)
-			.hasMessageContaining("query cannot be null");
-	}
-
-	@Test
-	void whenPromptHasMissingNumberPlaceholderThenThrow() {
-		PromptTemplate customPromptTemplate = new PromptTemplate("You are the boss. Original query: {query}");
-		assertThatThrownBy(() -> MultiQueryExpander.builder()
-			.chatClientBuilder(mock(ChatClient.Builder.class))
-			.promptTemplate(customPromptTemplate)
-			.build()).isInstanceOf(IllegalArgumentException.class)
-			.hasMessageContaining("The following placeholders must be present in the prompt template")
-			.hasMessageContaining("number");
-	}
-
-	@Test
-	void whenPromptHasMissingQueryPlaceholderThenThrow() {
-		PromptTemplate customPromptTemplate = new PromptTemplate("You are the boss. Number of queries: {number}");
-		assertThatThrownBy(() -> MultiQueryExpander.builder()
-			.chatClientBuilder(mock(ChatClient.Builder.class))
-			.promptTemplate(customPromptTemplate)
-			.build()).isInstanceOf(IllegalArgumentException.class)
-			.hasMessageContaining("The following placeholders must be present in the prompt template")
-			.hasMessageContaining("query");
-	}
-
-}
